STEP3 Add insider sentiment page
Add a new page to show insider sentimal
data.

Add routes to navigate between the main
page and the newly created page. The main
page is now called dashboard.

Moved company name handling to additional component.
Moved the display of an arrow depending of the change to a directive.
This avoids dupliacated implementations.